requirements.txt
telegraph
requests[socks]
pillow
shutilwhich
imageio
urllib3==1.25.11
six
windows-curses
